pyproject.toml
[tool.poetry]
name = "twofactorauthcli"
license = "MIT"
version = "0.1.2"
description = "2FA in your computer"
authors = ["Angel Fernandez <angelfernandezibanez@gmail.com>"]
readme = "README.md"
repository = "https://github.com/anxodio/twofactorauthcli"

[tool.poetry.dependencies]
python = "^3.6"
pyotp = "^2.3.0"
pyyaml = "^5.3"
tabulate = "^0.8.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^19.10b0", allow-prereleases = true}
flake8 = "^3.7.9"

[tool.poetry.scripts]
2fa = "twofactorauthcli:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
